Moved functions to other files for better structuring
